When copying a region, take the real region into account
Currently when copying a region, it will transform any type of region
into a plain cuboid. When using region like poly, this is mostly not
what people actually want. Thus we when copying the region, replace all
block outside the defined region with air, which can later be masked out
when pasting.