Fix ExceptionHandler compareTo method.
The previous implementation violated the contract that compareTo must be behave the same way when operands are reversed and that it is transitive. Any classes that were not equal, or a superclass of another were returned as being "less than".

The new implementation tests if classes are superclasses of each other, otherwise falls back to alphabetical sorting on canonical name.

While this did not manifest in 1.6, the replacement sort algorithm in 1.8 (TimSort) caused the sorting to appear random due to apparently contiguous sections being consumed, then sorted in bulk.