Fixes and made commit sound nicer
Sorry for sounding like a jerk before the threading part. This will still print 1+1 like normal, but the for loop will run behind a thread and may take 2-3 minutes to execute.

Imports have been fixed and should be fine now. I personally haven't touched that part and guess Intellij automatically set it to that.