Default VALIDATION to off
People have issues copy/pasting with this on, so I think for now it's
better to leave it off. If someone needs the special behavior for
setting panes or fences, they can turn it back on.

We do need to investigate the reason why it's not working, and
potentially save validation for post-edit, rather than prior to
placement.