Make BukkitWorld lazier for frequent use-case
The adapters are used quite frequently by plugins using the WorldEdit
API, with BukkitWorld being one of the more common ones among plugins
such as WorldGuard. As a new BukkitWorld may need to be initialized
every time a physics (redstone and others) event happens, there is a
significant cost to anything put inside the constructor for BukkitWorld.

This commit attempts to reduce the initialization cost by storing a
reference to the world's name instead of holding a WeakReference, which
may actually fix bugs relating to code assuming the world is still
loaded because the reference was not collected yet.

As well it lazily initializes the WorldNativeAccess when it's needed,
which is usually only for modifications that are actually applied to the
world. It does the same for the effect queue. These changes should be
beneficial to everyone, and significantly improve allocations.