Add utils to make working with tools and brushes easier.
Added methods to check if a item has a tool or brush tool bound.
Added method to get a optional tool depending on the class it is assignable to. This avoids existence and instance checks and gives clearer communitcation
Added method to retrieve a brush assigned to a brush tool if it matches a specific class