fix hang on shutdown for fabric due to hanging user threads
The JVM initiates shutdown when all non-daemon java threads have
terminated. When saving a schematic two threads are created that
are never stopped, thus preventing the server for finishing the
shutdown.

One is the "WorldEdit Task Executor - 0", shich is easily fixed.

The second is the TimerThread created by FutureProgressListener.
In this patch the timer is made to use a deamon thread instead. If
critical scheduled events on this timer persost after a server
shutdown is otherwise complete maybe this needs to be changed to
an executor service? Otherwise scheduled events may get lost by
the server shutting down.