Fix structure block CUI resetting selections cross-world.
Previously, if a world override was active (e.g. via //world), createStructureBlock would always attempt to check the player (entity) world rather than their selection world. This led to the selection getting reset immediately. Now, we first check if there's anything to render (i.e. if the selection is in the current world) before attempting to get the selection from the player's session.